Add negation to filename and fileext, use same syntax as with content.

+++ src/detect-filename.h
@@ -23,13 +23,14 @@
 
 #ifndef __DETECT_FILENAME_H__
 #define __DETECT_FILENAME_H__
+
 #include "util-spm-bm.h"
 
 typedef struct DetectFilenameData {
     uint8_t *name; /** name of the file to match */
     BmCtx *bm_ctx; /** BM context */
     uint16_t len; /** name length */
-    uint8_t flags;
+    uint32_t flags;
 } DetectFilenameData;
 
 /* prototypes */
